Adds bucket names
